+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - 优化 `f-back-top` 组件实现
+- `f-menu` 组件新增 `on-menu-item-click` `on-submenu-click` 配置项
 
 ## 0.25.1 (2023-02-21)